database done.

+++ home.php
@@ -1,10 +1,7 @@
 <?php  
 	session_start();
 
-	if (!isset($_SESSION['username'])) {
-		echo "logged out";
-		header('location:login.php');
-	}
+	
 ?>
 
 
@@ -16,10 +13,19 @@
 	<link rel="stylesheet" type="text/css" href="styles/styles.css">
 </head>
 <body>
-	<?php include 'nav.php';?>
-	<h1 class="m-auto">Home</h1>
+
 	
-	
+<?php include 'nav.php';?>
+<form action="home.php" method="post">
+<select name="category">
+	<option value="sports" >sports</option>
+	<option value="tech" >tech</option>
+	<option value="books" >books</option>
+
+</select>
+
+<button name="submit" class="btn btn-danger">filter</button>
+</form>
 
 
 <?php 
@@ -27,7 +33,18 @@ include 'connection.php';
 
 	$result_per_page =  4;
 	$sql = "SELECT * FROM post";
+	if (isset($_POST['submit'])) {
+	echo $filter = $_POST['category'];
 
+	$sql = "SELECT * FROM `post` WHERE category = '$filter'";
+}else{
+	$sql = "SELECT * FROM post";
+
+}
+	
+
+
+	
 	$result = mysqli_query($con,$sql);
 
 	$no_of_result = mysqli_num_rows($result);
@@ -35,7 +52,6 @@ include 'connection.php';
 
 
 	$no_of_pages = ceil($no_of_result/$result_per_page);
-
 
 	
 
@@ -47,9 +63,14 @@ if (!isset($_GET['page'])) {
 
 $this_page_first_result = ($page - 1)*$result_per_page; 
 
-$sql = "SELECT * FROM post LIMIT $this_page_first_result, 
+if (isset($_POST['submit'])) {
+	$filter = $_POST['category'];
+$sql = "SELECT * FROM post where category = '$filter' LIMIT $this_page_first_result, 
+$result_per_page ";
+}else{
+	$sql = "SELECT * FROM post LIMIT $this_page_first_result, 
 $result_per_page";
-
+}
 
 $result = mysqli_query($con,$sql);
 $rows  = mysqli_fetch_assoc($result);
@@ -68,9 +89,11 @@ while ($row  = mysqli_fetch_assoc($result) ) {?>
 		    <p class="card-text"><?php echo $row['content'] ;?></p>
 
 		    <small>Date Posted:<?php echo $row['date'];?></small>
+		    <small>category:<?php echo $row['category'];?></small><br>
+		  <?php  if (isset($_SESSION['username'])) {?>
 		    <a href="delete.php?id=<?php echo $row['id'];?>" class="btn btn-danger">Delete</a>
 		    <a href="update.php?id=<?php echo $row['id'];?>" class="btn btn-primary">Edit</a> 
-
+<?php }?>
 		    
 		  </div>
 	</div>
@@ -83,7 +106,7 @@ for($page = 1 ; $page<=$no_of_pages ; $page++){
  
   
 
-  <?php include 'footer.php'?>
+  
 
 <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
@@ -91,7 +114,7 @@ for($page = 1 ; $page<=$no_of_pages ; $page++){
   <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.16.0/umd/popper.min.js"></script>
   <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
-
+<?php include 'footer.php'; ?>
   
 </body>
 </html>
